+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -17,6 +17,13 @@
 * The caches used to resolve event full names are now bounded.
   ([#4857](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4857))
 
+## 1.18.0
+
+Released 2026-Aug-21
+
+* Updated OpenTelemetry core component version(s) to `1.18.0`.
+  ([#5022](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5022))
+
 ## 1.17.0
 
 Released 2026-Jul-17